requirements.txt
graphics.py>=5.0